remove confusing code
